backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-multipart>=0.0.6
openai>=1.0.0
chromadb>=0.4.0
pymupdf>=1.23.0
google-generativeai>=0.3.0
google-search-results>=2.4.0
requests>=2.31.0
python-dotenv>=1.0.0
